2_ml_pipeline_project/requirements.txt
dagster==1.10.4
dagster-webserver==1.10.4
dagster-airbyte==0.26.4
dagster-dbt==0.26.4
dagster-mlflow==0.26.4
dbt-core==1.5.2
protobuf==4.25.3
pendulum==2.1.2
sqlalchemy<2
psycopg2-binary
pandas
scikit-learn
requests
python-dotenv
